fix: resolve eslint errors - split filter logic from components
- Move useFilterState and applyFilters to hooks/use-filter-state.ts
  to satisfy react-refresh/only-export-components rule
- Remove typeof import() annotation in test to fix
  consistent-type-imports rule
