+++ frontend/src/features/cookbook/components/catalogue-grid.test.tsx
@@ -35,7 +35,7 @@ const mockItems: CookbookItem[] = [
 
 const mockNavigate = vi.fn()
 vi.mock('react-router-dom', async () => {
-  const actual = await vi.importActual<typeof import('react-router-dom')>('react-router-dom')
+  const actual = await vi.importActual('react-router-dom')
   return { ...actual, useNavigate: () => mockNavigate }
 })
